Fix key mapping type in Wayland window manager for improved compatibility
- Changed the type of the key symbol parameter in the input mapping function from uint64_t to std::uint32_t, ensuring proper handling of key codes.
- This adjustment enhances compatibility with the Wayland input system and maintains existing functionality.

+++ src/vertexnova/xwin/platform/linux/wayland/wayland_window_manager.cpp
@@ -375,7 +375,7 @@ void WaylandWindowManager::onKey(uint32_t linux_key, uint32_t state, uint32_t /*
     }
     const WindowDescriptor& desc = win->descriptor();
     const vne::events::KeyCode kc = mapNativeKeyToEvents(WindowAPI::eWaylandWindow,
-                                                         packXkbNativeKey(static_cast<uint64_t>(sym)),
+                                                         packXkbNativeKey(static_cast<std::uint32_t>(sym)),
                                                          desc.input_mapping.get());
     const uint8_t base_mods = mapWaylandModifiers(mod_depressed_, mod_latched_, mod_locked_);
     const uint8_t mods = mapNativeModifiersToEvents(WindowAPI::eWaylandWindow,
